+++ vicohome_bridge/vico-cli-main/cmd/events/list.go
@@ -71,12 +71,13 @@ Times should be in format: 2025-05-18 14:59:25`,
 
 		startTimestamp := fmt.Sprintf("%d", start.Unix())
 		endTimestamp := fmt.Sprintf("%d", end.Unix())
+		countryCode := auth.GetCountryCode()
 
 		eventsReq := Request{
 			StartTimestamp: startTimestamp,
 			EndTimestamp:   endTimestamp,
 			Language:       "en",
-			CountryNo:      "US",
+			CountryNo:      countryCode,
 		}
 
 		events, err := fetchEvents(token, eventsReq)